css/css-overflow/overflow-clip-scroll-size.html

<!doctype html>
<meta charset="utf-8">
<title>overflow: scroll width/height should return overflow size</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow/#valdef-overflow-clip">
<link rel="author" title="Scott Violet" href="mailto:sky@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  .parent {
    width: 100px;
    height: 101px;
  }
  .child {
    width: 200px;
    height: 201px;
  }
</style>
<div id="parent-clip-both" class="parent" style="overflow: clip">
  <div class="child"></div>
</div>
<div id="parent-clip-x" class="parent" style="overflow: clip-x">
  <div class="child"></div>
</div>
<div id="parent-clip-y" class="parent" style="overflow: clip-y">
  <div class="child"></div>
</div>
<script>
test(() => {
  var pClipBoth = document.getElementById("parent-clip-both");
  assert_equals(pClipBoth.scrollWidth, 200);
  assert_equals(pClipBoth.scrollHeight, 201);

  var pClipX = document.getElementById("parent-clip-x");
  assert_equals(pClipX.scrollWidth, 200);
  assert_equals(pClipX.scrollHeight, 201);

  var pClipY = document.getElementById("parent-clip-y");
  assert_equals(pClipY.scrollWidth, 200);
  assert_equals(pClipY.scrollHeight, 201);
}, "scroll size should match that of size specified by overflow: clip");
</script>
